+++ antlr/src/main/kotlin/com/strumenta/kolasu/antlr/parsing/testing.kt
@@ -1,4 +1,4 @@
-package com.strumenta.kolasu.parsing
+package com.strumenta.kolasu.antlr.parsing
 
 import org.antlr.v4.runtime.ParserRuleContext
 import org.antlr.v4.runtime.Vocabulary
