+++ antlr/src/test/kotlin/com/strumenta/kolasu/antlr/mapping/ParseTreeToASTTransformerAnotherTest.kt
@@ -1,4 +1,4 @@
-package com.strumenta.kolasu.mapping
+package com.strumenta.kolasu.antlr.mapping
 
 import com.strumenta.kolasu.model.Named
 import com.strumenta.kolasu.model.Node
