+++ antlr/src/test/kotlin/com/strumenta/kolasu/antlr/model/OriginTest.kt
@@ -1,6 +1,12 @@
-package com.strumenta.kolasu.model
+package com.strumenta.kolasu.antlr.model
 
-import com.strumenta.kolasu.parsing.ParseTreeOrigin
+import com.strumenta.kolasu.antlr.parsing.ParseTreeOrigin
+import com.strumenta.kolasu.model.Node
+import com.strumenta.kolasu.model.Point
+import com.strumenta.kolasu.model.Position
+import com.strumenta.kolasu.model.pos
+import com.strumenta.kolasu.model.withOrigin
+import com.strumenta.kolasu.model.withPosition
 import com.strumenta.simplelang.SimpleLangLexer
 import com.strumenta.simplelang.SimpleLangParser
 import org.antlr.v4.runtime.CharStreams
