+++ emf/src/test/kotlin/com/strumenta/kolasu/emf/ModelTest.kt
@@ -1,8 +1,8 @@
 package com.strumenta.kolasu.emf
 
+import com.strumenta.kolasu.antlr.parsing.withParseTreeNode
 import com.strumenta.kolasu.model.*
 import com.strumenta.kolasu.model.Statement
-import com.strumenta.kolasu.parsing.withParseTreeNode
 import com.strumenta.kolasu.transformation.GenericNode
 import com.strumenta.simplelang.SimpleLangLexer
 import com.strumenta.simplelang.SimpleLangParser
